separated scripts from modules into /scripts
