+++ src/rank_llm/scripts/run_response_analysis_verbose.py
@@ -1,0 +1,27 @@
+import argparse
+
+import sys
+import os
+
+SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
+parent = os.path.dirname(SCRIPT_DIR)
+parent = os.path.dirname(parent)
+sys.path.append(parent)
+
+from rank_llm.rerank.rankllm import PromptMode
+from rank_llm.analysis.response_analysis_verbose import ResponseAnalyzer
+
+def main(args):
+    response_analyzer = ResponseAnalyzer(args.files, 100, PromptMode.RANK_GPT)
+    responses, num_passages = response_analyzer.read_saved_responses()
+    print("Normalized scores:")
+
+    print(response_analyzer.count_errors(responses, num_passages, args.verbose))
+    # Print normalized scores
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--files", type=str, nargs="+", required=True)
+    parser.add_argument("--verbose", action="store_true")
+    args = parser.parse_args()
+    main(args)
